+++ CHANGELOG.md
@@ -2,20 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Support keyword `IDF_VERSION` in the if statement of the manifest file
+
+### Fixed
+
+- Build failed with warnings even without passing `--check-warnings`
+
 ## [1.0.1] (2023-06-12)
 
-## Fixed
+### Fixed
 
 - glob patterns are matched recursively
 
 ## [1.0.0] (2023-05-25)
 
-## Added
+### Added
 
 - Support keyword `depends_filepatterns` in the manifest file
 - Support expanding environment variables in the manifest files
 
-## BREAKING CHANGES
+### BREAKING CHANGES
 
 - Attributes Renamed
   - `App.requires_components` renamed to `App.depends_components`